requirements.txt
torch
transformers
pandas
scikit-learn
pdfminer.six
python-docx
streamlit
pymongo 
python-dotenv
sentencepiece